common: extract the model resolution into its own unit
