+++ common/CMakeLists.txt
@@ -84,6 +84,8 @@ add_library(${TARGET}
     llguidance.cpp
     log.cpp
     log.h
+    model-resolution.cpp
+    model-resolution.h
     ngram-cache.cpp
     ngram-cache.h
     ngram-map.cpp
