+++ common/download.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "hf-cache.h"
+#include "model-resolution.h"
 
 #include <string>
 #include <vector>
@@ -103,13 +104,4 @@ std::string common_docker_resolve_model(const std::string & docker);
 // returns true if anything was removed
 bool common_download_remove(const std::string & hf_repo_with_tag);
 
-struct common_download_hf_plan {
-    hf_cache::hf_file primary;
-    hf_cache::hf_files model_files;
-    hf_cache::hf_file mmproj;
-    hf_cache::hf_file mtp;
-    hf_cache::hf_file eagle3;
-    hf_cache::hf_file dflash;
-    hf_cache::hf_file preset; // if set, only this file is downloaded
-};
 common_download_hf_plan common_download_get_hf_plan(const common_params_model & model, const common_download_opts & opts);
